lessons/M3/MCP/module_3_mcp/requirements.txt
# requirement related to the mcp tutorial. If need to install ollama from ollama.com

ollama==0.5.4
langchain-ollama==0.3.8
# "mcp[cli]"
fastmcp==0.1.3
openai
beautifulsoup4
requests
